move PageLayout into app/components/layout folder
