+++ templates/skeleton/app/components/README.md
@@ -1,0 +1,4 @@
+layout define the overall structure of a page.
+
+Components in this folder includes HTML that wraps around all the pages (including default ErrorBoundary) in this application,
+header, footer, and where the main content area should go.
