+++ templates/skeleton/app/components/layout/Footer.tsx
